feat(cli): Extract linter commands to modular package (PR3)
Extract all 10 linter CLI commands from src/cli_main.py (1,239 lines) into
organized src/cli/linters/ package, reducing cli_main.py to 34 lines.

New modules created:
- src/cli/linters/shared.py - Common utilities for linter commands
- src/cli/linters/structure_quality.py - nesting, srp commands
- src/cli/linters/code_smells.py - dry, magic-numbers commands
- src/cli/linters/code_patterns.py - print-statements, method-property, stateless-class
- src/cli/linters/structure.py - file-placement, pipeline commands
- src/cli/linters/documentation.py - file-header command

Key changes:
- Fixed circular import in src/cli/__init__.py by importing from main.py directly
- Added __version__: str type annotation in src/__init__.py for MyPy
- Added E402 per-file-ignore for src/__init__.py in .flake8
- Added DRY ignores for CLI linter modules (framework adapter pattern)
- Full type annotations with TYPE_CHECKING guards for Orchestrator
